REFACTOR :: Icon들 ui 폴더로 이동
