+++ packages/ui/src/icon/LikeIcon.tsx
@@ -6,5 +6,4 @@ const LikeIcon = (props: SVGProps<SVGSVGElement>) => (
     </svg>
   );
     
-  export default LikeIcon;
-    
+export default LikeIcon;
