+++ packages/ui/src/icon/index.ts
@@ -1,0 +1,2 @@
+export { default as LikeIcon } from "./LikeIcon";
+export { default as SearchIcon } from "./SearchIcon";
