+++ packages/ui/src/index.ts
@@ -1,3 +1,4 @@
 // export * from "./assets"
 export * from "./components"
+export * from "./icon"
 export * from "./styles"
